chore: refactor components and build headlines for itinerary overviews
