+++ visual-debugger/src/data-processing/planParsing.ts
@@ -108,5 +108,3 @@ export function resetCssClassesForPlanEntries(plan: Plan) {
         itinerary.cssClass = itineraryStates.planEntryDefault
     }
 }
-
-
