+++ visual-debugger/src/lib/components/ui/plan-overviews/DefaultPlanOverview.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import {currentDefaultPlanStore} from "sveltestore";
     import type {Itinerary} from "@data/type-declarations/planTypes.ts";
-    import PlanEntry from "$lib/components/ui/subcomponents/PlanEntry.svelte";
+    import PlanEntry from "@/components/ui/subcomponents/PlanEntry.svelte";
     import {ScrollArea} from "@/components/ui/scroll-area";
 
     let itineraries: Itinerary[]
