+++ visual-debugger/src/lib/components/ui/subcomponents/ConnectionDetail.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
     import ArrowRight from 'lucide-svelte/icons/arrow-right';
-    import type {Itinerary, Leg} from '@data/type-declarations/planTypes';
+    import type {Itinerary, Leg} from '@data/type-declarations/planTypes.ts';
     import Time from '@/motis-helpers/Time.svelte';
     import {routeBorderColor, routeColor} from '@/motis-helpers/modeStyle.ts';
     import {formatDurationSec, formatDistanceMeters} from '@/motis-helpers/formatDuration.ts';
-    import {Button} from '$lib/components/ui/button';
+    import {Button} from '@/components/ui/button';
     import Route from '@/components/ui/Route.svelte';
     import {getModeName} from '@/motis-helpers/getModeName.ts';
-    import {t} from '@/motis-helpers/i18n/translation';
+    import {t} from '@/motis-helpers/i18n/translation.ts';
     import {type ItineraryShadow, LegShadow} from "@data/type-declarations/shadowTypes.ts";
     import MatchIndicator from "@/components/ui/subcomponents/MatchIndicator.svelte";
